backend/requirements.txt
langchain-mcp-adapters
langchain-pinecone
langchain-community
langchain-text-splitters
langchain-tavily
pypdf
fastapi
uvicorn
langchain-openai
ipykernel
bcrypt==5.0.0
certifi==2025.11.12
langgraph
passlib==1.7.4
python-dotenv
sqlmodel
python-jose[cryptography]